backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2
alembic==1.13.0
python-dotenv==1.0.0
apscheduler==3.10.4
